Add the package and library to an import diagnostic.
While most of the examples in our tests are made clear by the code
snippet, sometimes that's not the case. This makes it more clear when
the implicit `Main` package is used for example. I added this because
I got an error message without any source code to show in the snippet
(importing the prelude for example) and this makes the error much more
understandable.