Handle an invalid loc_id in SemIRLoc (#3890)
This is mainly to prevent a crash. We should typically have a valid
location, but in cases of bugs I think this is better than a simple
check-failure (namely, provides the diagnostic to give better context of
why the diagnostic location looks bad). I also think we might have more
invalid locations in the future.

Note I ran into this because we weren't setting the definition_id for a
class in import_ref; I think this made it much easier to examine.