Add a minimal SECURITY.md (#3885)
Also mentioned on DIscord
[here](https://discord.com/channels/655572317891461132/707150492370862090/1229512261832409238)
for context.

The template may seen be people with commit access at
https://github.com/carbon-language/carbon-lang/security/policy, by
clicking "Start setup". We don't have any releases so I'm eliding
support information there.

This stems from the security policy advice at
[OSSF](https://github.com/ossf/scorecard/blob/b118c1950f34e9d89237c1c6e69d273a192c490c/docs/checks.md#security-policy).
It's common enough for these to be minimal
([step-security](https://github.com/step-security/harden-runner/security)
is just an email address). I considered setting up a security list for
Carbon, but not sure we need it at the moment; it seems slightly
duplicative of the reporting which was previously enabled.