Stop skipping JSON copyright & license headers by default.
A lot of the JSON we use is parsed by VSCode or other systems that are
happy to have comments in them, and we should retain a license header
there. We can exclude specific files where this is a problem. I think
the `package.json` and generated lock file from it are two such cases
where we can't have the license header, but we can directly exclude more
as needed.

This should help make sure that when we *do* have a license header in
a JSON file it stays up to date and is formatted canonically.