Start building and bundling LLD for linking.
This is the first, minor steps in sketching out an installation tree for
the driver, and putting LLD into it in an LLVM-compatible location. It
then teaches the Clang runner to configure this private LLVM install in
the driver, and the link command to use LLD. With that, we use our LLD
rather than the system linker.

There are a couple of TODOs here that I plan to address in direct
follow-ups. First, I want to extract all the installation tree
management out from everything else and into a centralized place where
we can document the structure and provide APIs for navigating it. Next
I want to move the prelude over to use that infrastructure as well as
LLD. I left these as TODOs as I wanted to make sure I had the core logic
of finding LLD working first, and would prefer to checkpoint that rather
than re-doing it locally.