Introduce a hashtable key context.
This allows customizing both the hash and equality comparison. It also
does so in a way that enables stateful key contexts such as indices into
a specific vector. This works by requiring the context to be packaged on
every API accepting a key. When the context is stateless and can be
default constructed, this works via a default argument. When it is
stateful, the argument must be provided explicitly.

This allows us to add a stress test that forces colliding hashes in
egregious ways, as well as allowing the staeful hashing. There are also
a number of basic cleanups to code that were uncovered while working on
this.